Movins class LdMcSpecialUnitary to a new file and adding minor fixes
